requirements.txt
yfinance
prefect
dbt-postgres
psycopg2-binary
